Refactor login/signup, delete userform
