+++ src/components/SignIn/UserForm.css
@@ -14,7 +14,7 @@ input[type=text] {
   box-sizing: border-box;
 }
 
-input[type=submit] {
+input[type=submit], .userform-box button {
   width: 100%;
   background-color: #4CAF50;
   color: white;
@@ -23,6 +23,12 @@ input[type=submit] {
   border: none;
   border-radius: 4px;
   cursor: pointer;
+}
+
+
+.userform-box button {
+ width: 15rem; 
+ margin: 2rem;
 }
 
 input[type=submit]:hover {
