+++ src/components/SignIn/view.js
@@ -1,18 +1,22 @@
 import React from 'react'
+import { Link } from 'react-router-dom'
 
 export default function SignInForm(props) {
 
   return (
-    <form onSubmit={props.onSubmit}>
-      <label>
-        Name:
+    <div className="userform-box">
+      <form onSubmit={props.onSubmit}>
+        <label>
+          Name:
       <input type="text" name="name" onChange={props.onChange} value={props.values.name} placeholder={props.values.name} />
-      </label>
-      <label>
-        Password:
+        </label>
+        <label>
+          Password:
       <input type="text" name="password" onChange={props.onChange} value={props.values.password} placeholder={props.values.password} />
-      </label>
-      <input type="submit" value="Submit" />
-    </form>
+        </label>
+        <input type="submit" value="Submit" />
+      </form>
+      <Link to='/sign-up'><button type="button">signUp</button></Link>
+    </div>
   )
 }
