--- src/components/UserForm/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-
-const UserForm = () => {
-  return 'sign-up/ sign-in'
-
-}
-
-export default UserForm
